requirements.txt
fastapi==0.135.1
httpx==0.28.1
langchain==1.2.10
langchain-core==1.2.17
langchain-openai==1.1.10
openai==2.24.0
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
uvicorn==0.41.0
tiktoken==0.12.0
